+++ src/Viewer.h
@@ -22,7 +22,7 @@ namespace minity
 		~Viewer();
 
 		void display();
-		int m_explosion = 0;
+		float m_explosion = 0.0f;
 
 		GLFWwindow * window();
 		Scene* scene();
